requirements.txt
alembic
aiohttp
asyncpg
beautifulsoup4
celery
eventlet
fastapi
fastapi-cache2
fastapi_filter
uvicorn[standard]
pandas
pydantic
pydantic-settings
redis
sqlalchemy
xlrd
